requirements.txt
bitcoinlib==0.6.5
black==22.8.0
charset-normalizer==2.1.1
Django==4.0.7
django-cors-headers==3.13.0
django-debug-toolbar==3.7.0
django-timezone-field==5.0
djangorestframework==3.14.0
numpy==1.23.3
pika==1.3.1
psycopg2-binary==2.9.3
python-bitcoin-tools==0.2.3
python-bitcoinlib==0.11.2
python-decouple==3.6
redis==4.3.4
requests==2.28.1
web3==5.31.0
scrypt==0.8.20
pycryptodome==3.15.0
pymempool==0.0.4
websocket-client==1.6.1